fastapi-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
requests==2.31.0
PyYAML==6.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
cryptography==41.0.7
python-dotenv==1.0.0
aiofiles==23.2.1
structlog==23.2.0
